Refactored back-end

+++ server/src/routes/view-card-types/DeleteCardTypeRoute.js
@@ -12,22 +12,21 @@ class DeleteCardTypeRoute {
 	deleteCardTypeFromDatabase() {
 		this.app.delete("/card-types/:id", (req, res) => {
 			let cardTypeId = req.params.id;
-			let isCardType = 0;
-			let countCards = 0;
-			new sql.Request().query(`select count(*) as number from [dbo].[CardType] where [Id] like ${cardTypeId}`, (err, result) => {
-					isCardType = result.recordset[0].number;
-				});
-			new sql.Request().query(`select count(*) as number from [dbo].[Card] where [CardTypeId] like ${cardTypeId}`, (err, result) => {
-					countCards = result.recordset[0].number;
-				});
-			new sql.Request().query(`delete from [dbo].[Card] where [CardTypeId] like ${cardTypeId}`, (err, result) => {
-				
-				});
 			new sql.Request().query(`delete from [dbo].[CardType] where [Id] like ${cardTypeId}`, (err, result) => {
-				res.json({
-					isCardType,
-					countCards,
-				});
+				let isCardType = result.rowsAffected[0];
+				if (isCardType == 0) {
+					res.json({
+						isCardType
+					});
+				} else {
+					new sql.Request().query(`delete from [dbo].[Card] where [CardTypeId] like ${cardTypeId}`, (err, result) => {
+						let countCards = result.rowsAffected[0];
+						res.json({
+							isCardType,
+							countCards,
+						});
+					});
+				}
 			});
 		});
 	}
